rename output module to output
